+++ perl/FSM/Package/Spec.pm
@@ -10,12 +10,28 @@ sub new ($class, %args) {
     return bless {
         name => $args{name},
         symbols => $args{symbols},
-        raw_ast => $args{raw_ast},
+        raw_ast => _clone($args{raw_ast}),
     }, $class;
 }
 
 sub name ($self) { return $self->{name} }
 sub symbols ($self) { return $self->{symbols} }
-sub raw_ast ($self) { return $self->{raw_ast} }
+sub raw_ast ($self) { return _clone($self->{raw_ast}) }
+
+sub _clone ($value) {
+    return undef unless defined $value;
+
+    if (ref($value) eq 'HASH') {
+        return {
+            map { $_ => _clone($value->{$_}) } sort keys %$value
+        };
+    }
+
+    if (ref($value) eq 'ARRAY') {
+        return [ map { _clone($_) } @$value ];
+    }
+
+    return $value;
+}
 
 1;